+++ src/auth/components/SignupForm.tsx
@@ -1,8 +1,8 @@
-import { LabeledTextField } from "src/core/components/LabeledTextField"
-import { Form, FORM_ERROR } from "src/core/components/Form"
-import signup from "src/auth/mutations/signup"
-import { Signup } from "src/auth/schemas"
 import { useMutation } from "@blitzjs/rpc"
+import signup from "src/app/(auth)/mutations/signup"
+import { Signup } from "src/app/(auth)/schemas"
+import { Form, FORM_ERROR } from "src/core/components/Form"
+import { LabeledTextField } from "src/core/components/LabeledTextField"
 
 type SignupFormProps = {
   onSuccess?: () => void
@@ -12,7 +12,9 @@ export const SignupForm = (props: SignupFormProps) => {
   const [signupMutation] = useMutation(signup)
   return (
     <div className="w-full max-w-md rounded-lg bg-white p-8 shadow-md">
-      <h1 className="mb-4 text-2xl font-bold text-gray-800">Create an Account</h1>
+      <h1 className="mb-4 text-2xl font-bold text-gray-800">
+        Create an Account
+      </h1>
 
       <Form
         submitText="Create Account"
@@ -24,7 +26,10 @@ export const SignupForm = (props: SignupFormProps) => {
             await signupMutation(values)
             props.onSuccess?.()
           } catch (error: any) {
-            if (error.code === "P2002" && error.meta?.target?.includes("email")) {
+            if (
+              error.code === "P2002" &&
+              error.meta?.target?.includes("email")
+            ) {
               // This error comes from Prisma
               return { email: "This email is already being used" }
             } else {
@@ -34,7 +39,12 @@ export const SignupForm = (props: SignupFormProps) => {
         }}
       >
         <LabeledTextField name="email" label="Email" placeholder="Email" />
-        <LabeledTextField name="password" label="Password" placeholder="Password" type="password" />
+        <LabeledTextField
+          name="password"
+          label="Password"
+          placeholder="Password"
+          type="password"
+        />
       </Form>
     </div>
   )
